+++ R/add_absolute_targets.R
@@ -14,21 +14,19 @@ NULL
 #'
 #' @param x \code{\link{ConservationProblem-class}} object.
 #'
-#' @param targets \code{numeric} \code{vector}, \code{character},
-#'   \code{matrix}, or \code{\link{ZonesCharacter}} object with targets for
+#' @param targets \code{numeric} \code{vector}, \code{matrix}, or
+#'   \code{character} \code{vector} object with targets for
 #'   features. The correct argument for \code{targets} depends on multiple
 #'   factors:
 #'   \describe{
 #'     \item{\code{numeric}}{This type of argument can be a
 #'        \code{numeric} \code{vector} containing multiple values for each
 #'        feature. Additionally, for convenience,
-#'        this type of argument can be a single value to assign each feature
-#'        the same target. If the argument to \code{x} contains multiple zones,
-#'        each feature is assigned the same target in each zone. Thus
-#'        \code{numeric} arguments cannot be used to specify different targets
-#'        for different features in different zones.}
+#'        this type of argument can be a single value to assign the
+#'        same target to each feature. Note that this type of argument
+#'        cannot be used to specify targets for problems with multiple zones.}
 #'
-#'    \code{\code{matrix}}{This type of argument for \code{targets} can be
+#'    \item{\code{matrix}}{This type of argument for \code{targets} can be
 #'      used to set targets for each feature in each zone. Here, each
 #'      row corresponds to a different feature in argument to \code{x},
 #'      each column corresponds to a different zone in argument to \code{x},
@@ -38,16 +36,9 @@ NULL
 #'    \item{\code{character}}{This type of argument for \code{targets} can be
 #'       used to set the target for each feature using the names of fields
 #'       (columns) in the feature data associated with the argument to
-#'       \code{x}. If the argument to \code{x} contains multiple zones,
-#'        each feature is assigned the same target in each zone. Thus
-#'        as with \code{numeric} arguments to \code{target}, \code{character}
-#'        arguments cannot be used to specify different targets for different
-#'        features in different zones.}
-#'
-#'    \code{\code{\link{ZonesCharacter}}}{This type of argument for
-#'      \code{targets} can be to set targets for each feature in each zone
-#'      using the names of fields (columns) in the feature data associated
-#'      with the argument to \code{x}.}
+#'       \code{x}. This type of argument can only be used when the
+#'       feature data associated with \code{x} is a \code{data.frame}.
+#'       This argument must contain a field (column) name for each zone.}
 #'
 #'  }
 #'
@@ -79,10 +70,9 @@ NULL
 #'      box = FALSE)
 #' }
 #'
-#' @aliases add_absolute_targets-method add_absolute_targets,ConservationProblem,numeric-method add_absolute_targets,ConservationProblem,matrix-method add_absolute_targets,ConservationProblem,character-method add_absolute_targets,ConservationProblem,ZonesCharacter-method
+#' @aliases add_absolute_targets-method add_absolute_targets,ConservationProblem,numeric-method add_absolute_targets,ConservationProblem,matrix-method add_absolute_targets,ConservationProblem,character-method
 #'
 #' @name add_absolute_targets
-#' @docType methods
 NULL
 
 #' @name add_absolute_targets
@@ -101,7 +91,11 @@ methods::setMethod(
   "add_absolute_targets",
   methods::signature("ConservationProblem", "numeric"),
   function(x, targets, ...) {
-    add_absolute_targets(x, matrix(targets, ncol = 1))
+    assertthat::assert_that(inherits(x, "ConservationProblem"),
+                            x$number_of_zones() == 1,
+                            length(targets) %in% c(1, x$number_of_features()))
+    add_absolute_targets(x, matrix(targets, nrow = x$number_of_features(),
+                                   ncol = 1))
 })
 
 #' @name add_absolute_targets
@@ -117,15 +111,10 @@ methods::setMethod(
       inherits(targets, "matrix"), is.numeric(targets),
       isTRUE(all(is.finite(targets))),
       isTRUE(all(targets >= 0.0)),
-      isTRUE(length(targets) > 0))
-    # assert that targets are compatible with problem
-    if (nrow(targets) == 1 && ncol(targets) == 1) {
-      targets <- matrix(targets, nrow = x$number_of_features(),
-                        ncol = x$number_of_zones())
-    }
-    assertthat::assert_that(
+      isTRUE(length(targets) > 0),
       nrow(targets) == x$number_of_features(),
       ncol(targets) == x$number_of_zones(),
+      all(targets >= 0),
       isTRUE(all(targets <= x$feature_abundances_in_planning_units())))
     # create targets as data.frame
     if (x$number_of_zones() > 1) {
@@ -136,7 +125,7 @@ methods::setMethod(
       target_data <- expand.grid(feature = x$feature_names(),
                                  type = "absolute")
     }
-    target_data$target <- c(targets)
+    target_data$target <- as.numeric(targets)
     # add targets to problem
     add_manual_targets(x, target_data)
   })
@@ -150,31 +139,13 @@ methods::setMethod(
   function(x, targets, ...) {
     # assert that arguments are valid
     assertthat::assert_that(inherits(x, "ConservationProblem"),
-      inherits(x$data$features, c("data.frame", "Spatial")),
-      assertthat::has_name(x$data$features, targets),
-      length(targets) == x$data$features())
+      is.character(targets),
+      inherits(x$data$features, "data.frame"),
+      !anyNA(targets),
+      all(assertthat::has_name(x$data$features, targets)),
+      length(targets) == x$number_of_zones(),
+      all(vapply(x$data$features[, targets, drop = FALSE], is.numeric,
+                logical(1))))
     # add targets to problem
-    add_absolute_targets(x, zones(targets))
- })
-
-#' @name add_absolute_targets
-#' @rdname add_absolute_targets
-#' @usage add_absolute_targets(x, targets, ...) # x=ConservationProblem, targets=ZonesCharacter
-methods::setMethod(
-  "add_absolute_targets",
-  methods::signature("ConservationProblem", "ZonesCharacter"),
-  function(x, targets, ...) {
-    # assert that arguments are valid
-    assertthat::assert_that(inherits(x, "ConservationProblem"),
-     inherits(x$data$features, c("data.frame", "Spatial")),
-     n_zone(targets) == x$number_of_zones(),
-     n_feature(targets) == x$number_of_features(),
-     assertthat::has_name(x$data$features, unlist(as.list(targets))),
-     all(vapply(x$data$features[, unlist(as.list(targets)), drop = FALSE],
-                is.numeric, logical(1))))
-    # extract target data
-    targets <- x$data$features[, unlist(as.list(targets)), drop = FALSE]
-    targets <- as.matrix(targets)
-    # add targets to problem
-    add_absolute_targets(x, targets)
+    add_absolute_targets(x, as.matrix(x$data$features[, targets, drop = FALSE]))
 })