+++ R/add_manual_targets.R
@@ -2,7 +2,7 @@
 NULL
 
 #' @export
-methods::setOldClass("tibble")
+methods::setOldClass("tbl_df")
 
 #' Add Manual Targets
 #'
@@ -43,6 +43,8 @@ methods::setOldClass("tibble")
 #'
 #'  }
 #'
+#' @param ... not used.
+#'
 #' @details
 #' Targets are used to specify the minimum amount or proportion of a feature's
 #' distribution that needs to be protected. Most conservation planning problems
@@ -57,7 +59,7 @@ methods::setOldClass("tibble")
 #'
 #' @seealso \code{\link{targets}}.
 #'
-#' @aliases add_manual_targets-method add_manual_targets,ConservationProblem,data.frame-method add_manual_targets,ConservationProblem,tibble-method
+#' @aliases add_manual_targets-method add_manual_targets,ConservationProblem,data.frame-method add_manual_targets,ConservationProblem,tbl_df-method
 #'
 #' @name add_manual_targets
 #'
@@ -80,41 +82,42 @@ methods::setMethod(
   "add_manual_targets",
   methods::signature("ConservationProblem", "data.frame"),
   function(x, targets, ...) {
-    add_manual_targets(x, tibble::as.tibble(x))
+    add_manual_targets(x, tibble::as.tibble(targets))
 })
 
 #' @name add_manual_targets
 #' @rdname add_manual_targets
-#' @usage \S4method{add_manual_targets}{ConservationProblem,tibble}(x, targets, ...)
+#' @usage \S4method{add_manual_targets}{ConservationProblem,tbl_df}(x, targets, ...)
 methods::setMethod(
   "add_manual_targets",
-  methods::signature("ConservationProblem", "tibble"),
+  methods::signature("ConservationProblem", "tbl_df"),
   function(x, targets, ...) {
     # assert that arguments are valid
-    validate_targets <- function(x) {
+    assertthat::assert_that(inherits(x, "ConservationProblem"))
+    validate_targets <- function(targets) {
       assertthat::assert_that(
-        inherits(x, "ConservationProblem"),
-        inherits(targets, "tibble"),
-        assert_that::has_name(targets, "feature"),
-        assert_that::has_name(targets, "target"),
-        assert_that::has_name(targets, "type"),
+        inherits(targets, "tbl_df"),
+        assertthat::has_name(targets, "feature"),
+        assertthat::has_name(targets, "target"),
+        assertthat::has_name(targets, "type"),
         is.character(targets$feature) || is.factor(targets$feature),
         all(as.character(targets$feature) %in% x$feature_names()),
-        is.numeric(targets$target),
+        is.numeric(targets$target), all(is.finite(targets$target)),
         is.character(targets$type) || is.factor(targets$type),
         all(targets$type %in% c("absolute", "relative")))
-      if (x$number_of_zones() > 1 || assert_that::has_name(targets, "zone")) {
+      if (x$number_of_zones() > 1 || assertthat::has_name(targets, "zone")) {
         assertthat::assert_that(
-          assert_that::has_name(targets, "zone"),
+          assertthat::has_name(targets, "zone"),
           is.character(targets$zone) || is.factor(targets$zone) ||
-          is.list(targets$zone))
+          is.list(targets$zone),
+          all(unlist(targets$zone) %in% x$zone_names()))
         assertthat::assert_that(
           all(vapply(targets$zone, inherits, logical(1),
               c("character", "factor"))),
           msg = paste("argument to targets has list-type cells in the column",
                       "\"zone\" which contain invalid values"))
       }
-      if (assert_that::has_name(targets, "sense"))
+      if (assertthat::has_name(targets, "sense"))
        assertthat::assert_that(
          is.character(targets$sense) || is.factor(targets$sense),
          all(as.character(targets$sense) %in% c(">=", "<=", "=")))
@@ -122,8 +125,8 @@ methods::setMethod(
     }
     validate_targets(targets)
     # define function to validate changes to the targets object
-    vfun <- function(x) inherits(try(validate_targets(x), silent = TRUE),
-                                class = "try-error")
+    vfun <- function(x) !inherits(try(validate_targets(x), silent = TRUE),
+                                  "try-error")
     # define function to render targets object
     rfun <- function(x)
       getFromNamespace("rhandsontable", "rHandsontableOutput")(x)
@@ -134,8 +137,8 @@ methods::setMethod(
     name = "Targets",
     data = list(abundances = x$feature_abundances_in_planning_units()),
     parameters = parameters(misc_parameter("Targets", targets, vfun, rfun)),
-    repr = function() {
-      targets <- self$parameters$get("targets")
+    repr = function(self) {
+      targets <- self$parameters$get("Targets")
       if (all(as.character(targets$type) == "relative")) {
         out <- "Relative"
       } else if (all(as.character(targets$type) == "absolute")) {
@@ -143,35 +146,42 @@ methods::setMethod(
       } else {
         out <- "Mixed"
       }
-      out <- paste0(" targets [targets (min: ", min(targets$target),
+      out <- paste0(out, " targets [targets (min: ", min(targets$target),
                     ", max: ", max(targets$target), ")]")
       return(out)
      },
      output = function(self) {
        # get data
-       targets <- self$parameters$get("targets")
+       targets <- self$parameters$get("Targets")
        abundances <- self$data$abundances
        # add zone column if missing
-       if (!assert_that::has_name(targets, "zone"))
-         targets$zone <- "zone"
+       if (!assertthat::has_name(targets, "zone"))
+         targets$zone <- colnames(abundances)[[1]]
        # convert zone column to list of characters if needed
-       if (!inherits(target$zone, "list"))
+       if (!inherits(targets$zone, "list"))
         targets$zone <- as.list(targets$zone)
        # add sense column if missing
-       if (!assert_that::has_name(targets, "sense"))
-        targets$sense <- ">="
-       # add compute relative targets as absolute targets
-       targets$value <- targets$target
+       if (!assertthat::has_name(targets, "sense"))
+         targets$sense <- ">="
+       targets$sense <- as.character(targets$sense)
+       # convert feature names to indices
+       targets$feature <- match(targets$feature, rownames(abundances))
+       # convert zone names to indices
+       for (i in seq_len(nrow(targets))) {
+         targets$zone[[i]] <- match(targets$zone[[i]], colnames(abundances))
+       }
+       # add compute relative targets as absolute targets and assign
+       # zone ids
+       targets$value <- as.numeric(targets$target)
        relative_rows <- which(targets$type == "relative")
-       feature_id <- match(targets$feature, rownames(abundances))
        for (i in seq_along(relative_rows)) {
-          zone_names <- targets$zone[[relative_rows[[i]]]]
-          zone_id <- match(zone_names, colnames(abundances))
-          abund_mtx <- as.matrix(data.frame(feature_id[i], zone_id))
+          zone_id <- targets$zone[[relative_rows[[i]]]]
+          feature_id <- targets$feature[[relative_rows[[i]]]]
+          abund_mtx <- as.matrix(data.frame(feature_id, zone_id))
           targets$value[relative_rows[i]] <- sum(abundances[abund_mtx]) *
                                              targets$target[relative_rows[i]]
        }
        # return tibble
-       return(targets)
+       return(targets[, c("feature", "zone", "sense", "value")])
      }))
 })