+++ R/parameters.R
@@ -328,8 +328,8 @@ numeric_parameter_array <- function(name, value, label,
 #'
 #' @param validator \code{function} to validate changes to the parameter. This
 #'   function must have a single argument and return either \code{TRUE} or
-#'   \code{FALSE} to if the a \code{\code[tibble]{tibble}} object passed to
-#'   the argument is valid.
+#'   \code{FALSE} depending on if the argument is valid candidate for the
+#'   parameter.
 #'
 #' @param widget \code{function} to render a \code{shiny} widget. This function
 #'   should must have a single argument that accepts a valid object and return